+++ client/candidateList/configurationProvider.js
