+++ client/candidateList/index.js
@@ -2,22 +2,22 @@ import angular from 'angular';
 import 'angular-material/angular-material.css';
 import routes from './routes';
 
-angular.module('prospectList', [
+angular.module('irCandidateList', [
   require('angular-aria'),
   require('angular-animate'),
   require('angular-sanitize'),
   require('angular-material'),
   require('angular-route')
 ])
-.controller('prospectListController', require('./controller'))
-.provider('prospectListConfig', require('./configurationProvider'))
+.controller('candidateListController', require('./controller'))
+.provider('candidateListConfig', require('./configurationProvider'))
 
-.config(function($routeProvider, prospectListConfigProvider) {
-    const routePrefix = prospectListConfigProvider.routePrefix;
+.config(function($routeProvider, candidateListConfigProvider) {
+    const routePrefix = candidateListConfigProvider.routePrefix;
 
     routes.forEach((route) => {
       $routeProvider.when(routePrefix + route.path, route);
     });
 });
 
-module.exports = 'prospectList';
+module.exports = 'irCandidateList';
