+++ client/candidateList/routes.js
@@ -3,7 +3,7 @@ module.exports = [
   {
     path: '/',
     templateUrl: require('./template.html'),
-    controller: 'prospectListController',
+    controller: 'candidateListController',
     controllerAs: 'pl',
     resolve: []
   }
