+++ client/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
-<html ng-app="irProspects">
+<html ng-app="irVivace">
   <head>
-    <title>InRhythm - Prospects</title>
+    <title>InRhythm - Vivace</title>
   </head>
   <body>
     <ng-view/>
