+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "prospect-list",
+  "name": "vivace",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
