+++ test/client/candidateList/controller_test.js
@@ -1,11 +1,11 @@
 const chai =  require('chai');
 
 
-describe('the prospectList controller', () => {
+describe('the candidateList controller', () => {
   let ctrl;
-  beforeEach(angular.mock.module('prospectList'));
+  beforeEach(angular.mock.module('irCandidateList'));
   beforeEach(angular.mock.inject($controller => {
-    ctrl = $controller(require('../../../client/prospectList/controller'));
+    ctrl = $controller(require('../../../client/candidateList/controller'));
   }));
 
   it('should not be null', () => {
